Renamed cli.h into cli.hpp, made cli and compiler more consistent
